+++ src/jsonLanguageService.ts
@@ -24,7 +24,7 @@ import {
 	ASTNode,
 	Color, ColorInformation, ColorPresentation,
 	LanguageServiceParams, LanguageSettings, DocumentLanguageSettings,
-	FoldingRange, JSONSchema, SelectionRange
+	FoldingRange, JSONSchema, SelectionRange, FoldingRangesContext, DocumentSymbolsContext, ColorInformationContext as DocumentColorsContext
 } from './jsonLanguageTypes';
 
 export type JSONDocument = {};
@@ -43,15 +43,15 @@ export interface LanguageService {
 	resetSchema(uri: string): boolean;
 	doResolve(item: CompletionItem): Thenable<CompletionItem>;
 	doComplete(document: TextDocument, position: Position, doc: JSONDocument): Thenable<CompletionList | null>;
-	findDocumentSymbols(document: TextDocument, doc: JSONDocument, context?: { resultLimit?: number }): SymbolInformation[];
-	findDocumentSymbols2(document: TextDocument, doc: JSONDocument, context?: { resultLimit?: number }): DocumentSymbol[];
+	findDocumentSymbols(document: TextDocument, doc: JSONDocument, context?: DocumentSymbolsContext): SymbolInformation[];
+	findDocumentSymbols2(document: TextDocument, doc: JSONDocument, context?: DocumentSymbolsContext): DocumentSymbol[];
 	/** deprecated, use findDocumentColors instead */
 	findColorSymbols(document: TextDocument, doc: JSONDocument): Thenable<Range[]>;
-	findDocumentColors(document: TextDocument, doc: JSONDocument, context?: { resultLimit?: number }): Thenable<ColorInformation[]>;
+	findDocumentColors(document: TextDocument, doc: JSONDocument, context?: DocumentColorsContext): Thenable<ColorInformation[]>;
 	getColorPresentations(document: TextDocument, doc: JSONDocument, color: Color, range: Range): ColorPresentation[];
 	doHover(document: TextDocument, position: Position, doc: JSONDocument): Thenable<Hover | null>;
 	format(document: TextDocument, range: Range, options: FormattingOptions): TextEdit[];
-	getFoldingRanges(document: TextDocument, context?: { rangeLimit?: number }): FoldingRange[];
+	getFoldingRanges(document: TextDocument, context?: FoldingRangesContext): FoldingRange[];
 	getSelectionRanges(document: TextDocument, positions: Position[], doc: JSONDocument): SelectionRange[];
 }