+++ src/services/jsonDocumentSymbols.ts
@@ -8,7 +8,7 @@ import * as Strings from '../utils/strings';
 import { colorFromHex } from '../utils/colors';
 
 import { SymbolInformation, SymbolKind, TextDocument, Range, Location, TextEdit, DocumentSymbol } from 'vscode-languageserver-types';
-import { Thenable, ColorInformation, ColorPresentation, Color, ASTNode, PropertyASTNode } from "../jsonLanguageTypes";
+import { Thenable, ColorInformation, ColorPresentation, Color, ASTNode, PropertyASTNode, DocumentSymbolsContext } from "../jsonLanguageTypes";
 
 import { IJSONSchemaService } from "./jsonSchemaService";
 
@@ -17,52 +17,51 @@ export class JSONDocumentSymbols {
 	constructor(private schemaService: IJSONSchemaService) {
 	}
 
-	public findDocumentSymbols(document: TextDocument, doc: Parser.JSONDocument, context?: { resultLimit?: number }): SymbolInformation[] {
+	public findDocumentSymbols(document: TextDocument, doc: Parser.JSONDocument, context: DocumentSymbolsContext= { resultLimit: Number.MAX_VALUE }): SymbolInformation[] {
 
 		let root = doc.root;
 		if (!root) {
 			return null;
 		}
 
-		let limit = context && typeof context.resultLimit === 'number' ? context.resultLimit : Number.MAX_VALUE;
+		let limit = context.resultLimit;
 
 		// special handling for key bindings
 		let resourceString = document.uri;
 		if ((resourceString === 'vscode://defaultsettings/keybindings.json') || Strings.endsWith(resourceString.toLowerCase(), '/user/keybindings.json')) {
 			if (root.type === 'array') {
 				let result: SymbolInformation[] = [];
-				root.items.forEach(item => {
-					if (limit > 0 && item.type === 'object') {
+				for (let item of root.items) {
+					if (item.type === 'object') {
 						for (let property of item.properties) {
-							if (limit > 0 && property.keyNode.value === 'key') {
-								if (property.valueNode) {
-									if (property.valueNode) {
-										limit--;
-										let location = Location.create(document.uri, getRange(document, item));
-										result.push({ name: Parser.getNodeValue(property.valueNode), kind: SymbolKind.Function, location: location });
-									}
-									return;
+							if (property.keyNode.value === 'key' && property.valueNode) {
+								let location = Location.create(document.uri, getRange(document, item));
+								result.push({ name: Parser.getNodeValue(property.valueNode), kind: SymbolKind.Function, location: location });
+								limit--;
+								if (limit <= 0) {
+									context.resultLimitExceeded = true;
+									return result;
 								}
 							}
 						}
 					}
-				});
+				}
 				return result;
 			}
 		}
 
 		let collectOutlineEntries = (result: SymbolInformation[], node: ASTNode, containerName: string): SymbolInformation[] => {
-			if (limit > 0) {
+			if (limit >= 0) {
 				if (node.type === 'array') {
 					node.items.forEach(node => collectOutlineEntries(result, node, containerName));
 				} else if (node.type === 'object') {
 					node.properties.forEach((property: PropertyASTNode) => {
-						if (limit > 0) {
-							limit--;
-							let location = Location.create(document.uri, getRange(document, property));
+						if (limit >= 0) {
 							let valueNode = property.valueNode;
 							if (valueNode) {
-								let childContainerName = containerName ? containerName + '.' + property.keyNode.value : property.keyNode.value;
+								limit--;
+								const location = Location.create(document.uri, getRange(document, property));
+								const childContainerName = containerName ? containerName + '.' + property.keyNode.value : property.keyNode.value;
 								result.push({ name: this.getKeyLabel(property), kind: this.getSymbolKind(valueNode.type), location: location, containerName: containerName });
 								collectOutlineEntries(result, valueNode, childContainerName);
 							}
@@ -73,47 +72,52 @@ export class JSONDocumentSymbols {
 			return result;
 		};
 		let result = collectOutlineEntries([], root, void 0);
+		if (limit < 0) {
+			result.pop();
+			context.resultLimitExceeded = true;
+		}
 		return result;
 	}
 
-	public findDocumentSymbols2(document: TextDocument, doc: Parser.JSONDocument, context?: { resultLimit?: number }): DocumentSymbol[] {
+	public findDocumentSymbols2(document: TextDocument, doc: Parser.JSONDocument, context: DocumentSymbolsContext = { resultLimit: Number.MAX_VALUE }): DocumentSymbol[] {
 
 		let root = doc.root;
 		if (!root) {
 			return null;
 		}
 
-		let limit = context && typeof context.resultLimit === 'number' ? context.resultLimit : Number.MAX_VALUE;
+		let limit = context.resultLimit;
 
 		// special handling for key bindings
 		let resourceString = document.uri;
 		if ((resourceString === 'vscode://defaultsettings/keybindings.json') || Strings.endsWith(resourceString.toLowerCase(), '/user/keybindings.json')) {
 			if (root.type === 'array') {
 				let result: DocumentSymbol[] = [];
-				root.items.forEach(item => {
-					if (limit > 0 && item.type === 'object') {
+				for (let item of root.items) {
+					if (item.type === 'object') {
 						for (let property of item.properties) {
-							if (limit > 0 && property.keyNode.value === 'key') {
-								if (property.valueNode) {
-									limit--;
-									let range = getRange(document, item);
-									let selectionRange = getRange(document, property.keyNode);
-									result.push({ name: Parser.getNodeValue(property.valueNode), kind: SymbolKind.Function, range, selectionRange });
+							if (property.keyNode.value === 'key' && property.valueNode) {
+								let range = getRange(document, item);
+								let selectionRange = getRange(document, property.keyNode);
+								result.push({ name: Parser.getNodeValue(property.valueNode), kind: SymbolKind.Function, range, selectionRange });
+								limit--;
+								if (limit <= 0) {
+									context.resultLimitExceeded = true;
+									return result;
 								}
-								return;
 							}
 						}
 					}
-				});
+				}
 				return result;
 			}
 		}
 
 		let collectOutlineEntries = (result: DocumentSymbol[], node: ASTNode): DocumentSymbol[] => {
-			if (limit > 0) {
+			if (limit >= 0) { // collect one more to know that limit got exceeded
 				if (node.type === 'array') {
 					node.items.forEach((node, index) => {
-						if (node && limit > 0) {
+						if (node && limit >= 0) {
 							limit--;
 							let range = getRange(document, node);
 							let selectionRange = range;
@@ -125,13 +129,12 @@ export class JSONDocumentSymbols {
 				} else if (node.type === 'object') {
 					node.properties.forEach((property: PropertyASTNode) => {
 						let valueNode = property.valueNode;
-						if (valueNode && limit > 0) {
+						if (valueNode && limit >= 0) {
 							limit--;
 							let range = getRange(document, property);
 							let selectionRange = getRange(document, property.keyNode);
 							let children = collectOutlineEntries([], valueNode);
 							result.push({ name: this.getKeyLabel(property), kind: this.getSymbolKind(valueNode.type), range, selectionRange, children });
-
 						}
 					});
 				}
@@ -139,6 +142,10 @@ export class JSONDocumentSymbols {
 			return result;
 		};
 		let result = collectOutlineEntries([], root);
+		if (limit < 0) {
+			result.pop();
+			context.resultLimitExceeded = true;
+		}
 		return result;
 	}
 
@@ -171,7 +178,7 @@ export class JSONDocumentSymbols {
 		return `"${name}"`;
 	}
 
-	public findDocumentColors(document: TextDocument, doc: Parser.JSONDocument, context?: { resultLimit?: number }): Thenable<ColorInformation[]> {
+	public findDocumentColors(document: TextDocument, doc: Parser.JSONDocument, context?: DocumentSymbolsContext): Thenable<ColorInformation[]> {
 		return this.schemaService.getSchemaForResource(document.uri, doc).then(schema => {
 			let result: ColorInformation[] = [];
 			if (schema) {
@@ -179,7 +186,7 @@ export class JSONDocumentSymbols {
 				let matchingSchemas = doc.getMatchingSchemas(schema.schema);
 				let visitedNode = {};
 				for (let s of matchingSchemas) {
-					if (limit > 0 && !s.inverted && s.schema && (s.schema.format === 'color' || s.schema.format === 'color-hex') && s.node && s.node.type === 'string') {
+					if (!s.inverted && s.schema && (s.schema.format === 'color' || s.schema.format === 'color-hex') && s.node && s.node.type === 'string') {
 						let nodeId = String(s.node.offset);
 						if (!visitedNode[nodeId]) {
 							let color = colorFromHex(Parser.getNodeValue(s.node));
@@ -189,6 +196,12 @@ export class JSONDocumentSymbols {
 							}
 							visitedNode[nodeId] = true;
 							limit--;
+							if (limit <= 0) {
+								if (context) {
+									context.resultLimitExceeded = true;
+								}
+								return result;
+							}
 						}
 					}
 				}